Actualizaciones para conexion con front

+++ src/main/java/munipalmares/sgd/controlador/denunciaControlador.java
@@ -26,13 +26,31 @@ public class denunciaControlador {
         return ResponseEntity.ok(denunciaRepositorio.save(denuncia));
     }
 
+//    @PostMapping("/denuncia")
+//    ResponseEntity<denuncia> saveDenuncia(@RequestBody denuncia den) {
+//
+//        // Verificar si se envió evidenciaFotografica como String Base64
+//        // Esto solo aplica si en Postman envías la imagen como String
+//        if (den.getEvidenciaFotografica() != null) {
+//            try {
+//                // Intentar decodificar Base64
+//                byte[] decoded = java.util.Base64.getDecoder().decode(new String(den.getEvidenciaFotografica()));
+//                den.setEvidenciaFotografica(decoded);
+//            } catch (IllegalArgumentException e) {
+//                // Si no es Base64, lo dejamos como está (ya es byte[])
+//            }
+//        }
+//
+//        return ResponseEntity.ok(denunciaRepositorio.save(den));
+//    }
+
     @PutMapping("/denuncia")
     ResponseEntity<denuncia> updateDenuncia(@RequestBody  denuncia denuncia) {
         return ResponseEntity.ok(denunciaRepositorio.save(denuncia));
     }
 
     @DeleteMapping("/denuncia/{id}")
-    ResponseEntity<Void> deleteDenuncia(@PathVariable("id") Long idDenuncia ){
+    ResponseEntity<Void> deleteDenuncia(@PathVariable("id") Integer idDenuncia ){
         denunciaRepositorio.deleteById(idDenuncia);
         return ResponseEntity.ok().build();
     }
